Store the index in a std::size_t

+++ Sprocket/Utility/SparseSet.h
@@ -11,7 +11,7 @@
 
 namespace Sprocket {
 
-template <typename ValueType, typename IndexType = std::uint16_t>
+template <typename ValueType, typename IndexType = std::size_t>
 class SparseSet
 {
 public:
